+++ api/handlers/config.go
@@ -245,14 +245,7 @@ func NewStorageService(providerType string) (*s3.S3, error) {
 		return s3.New(sess), nil
 
 	case "aws-s3":
-		region := os.Getenv("AWS_REGION")
-		accessKeyID := os.Getenv("AWS_ACCESS_KEY_ID")
-		secretAccessKey := os.Getenv("AWS_SECRET_ACCESS_KEY")
-
-		sess, err := session.NewSession(&aws.Config{
-			Region:      aws.String(region),
-			Credentials: credentials.NewStaticCredentials(accessKeyID, secretAccessKey, ""),
-		})
+		sess, err := session.NewSession()
 		if err != nil {
 			return nil, fmt.Errorf("error creating s3 session: %s", err.Error())
 		}